MindMate-AI/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
pypdf>=3.17.0
ollama>=0.1.5
openai>=1.12.0
llama-index>=0.10.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.1
pytube>=15.0.0